requirements.txt
loguru
thop
tensorboard
torchsummary
jieba
tabulate
tqdm
boto3
scipy
ujson
sentencepiece
regex
scikit-learn
psutil
seqeval
packaging
sacrebleu
opencc
einops
rouge